Gateway&Traffic Limit&GRPC&JVM Optimization (#53)
**1. Gateway**
There're memory leak in the SGC.
The SGC Performance and scalability require verification.
The traffic limit is required in case of DDoS and pressure test via SGC
or Nginx.
Allow&Deny Policy is required for some IP or IP subnets.
**2. GRPC**
Find a suitable gPRC scenario for the serverless.
**3. JVM Optimization**
a. Find the appropriate GC
b. Resolve the memory leak.